+++ lib/render/mayaToHydra/utils.h
@@ -37,21 +37,18 @@ struct MtohRendererDescription {
 };
 
 using MtohRendererDescriptionVector = std::vector<MtohRendererDescription>;
-using MtohRendererSettingsVector = std::vector<HdRenderSettingDescriptorList>;
+
+// Map from MtohRendererDescription::rendererName to it's a HdRenderSettingDescriptorList
+using MtohRendererSettings =
+    std::unordered_map<TfToken, HdRenderSettingDescriptorList,
+                       TfToken::HashFunctor>;
 
 TfTokenVector MtohGetRendererPlugins();
 std::string MtohGetRendererPluginDisplayName(const TfToken& id);
 TfToken MtohGetDefaultRenderer();
 const MtohRendererDescriptionVector& MtohGetRendererDescriptions();
+const MtohRendererSettings& MtohGetRendererSettings();
 
-// Initialize the render-delegates list to all known valid delegates.
-// This function should only be called once (on start-up).
-//
-// The second item (MtohRendererSettingsVector) is only valid from that call
-// as continuing initialization may relocate/move the items.
-//
-using MtohRendererInitialization = std::pair<const MtohRendererDescriptionVector&, MtohRendererSettingsVector>;
-MtohRendererInitialization MtohInitializeRenderPlugins();
 
 PXR_NAMESPACE_CLOSE_SCOPE
